Docker-compose/32-crud/requirements.txt
django
djangorestframework
psycopg2-binary
django_filter
gunicorn
python-decouple
